Move storm-dft/modelchecker/dft storm-dft/modelchecker
