+++ src/storm-dft/utility/FailureBoundFinder.h
@@ -1,5 +1,5 @@
 #include <vector>
-#include "storm-dft/modelchecker/dft/DFTASFChecker.h"
+#include "storm-dft/modelchecker/DFTASFChecker.h"
 #include "storm-dft/storage/dft/DFT.h"
 
 namespace storm {
