+++ src/test/storm-dft/bdd/TestBdd.cpp
@@ -5,8 +5,8 @@
 
 #include "storm-config.h"
 #include "storm-dft/api/storm-dft.h"
-#include "storm-dft/modelchecker/dft/SFTBDDChecker.h"
-#include "storm-dft/modelchecker/dft/SFTBDDPropertyFormulaAdapter.h"
+#include "storm-dft/modelchecker/SFTBDDChecker.h"
+#include "storm-dft/modelchecker/SFTBDDPropertyFormulaAdapter.h"
 #include "storm-dft/transformations/SftToBddTransformator.h"
 #include "storm-dft/utility/MTTFHelper.h"
 #include "storm-parsers/api/properties.h"
